Create server_prompt_cache.diff
Fixes problem in llama.cpp server where results will change depending if prompt cache is turned on or off.  Ensures identical numerical results with prompt cache on or off while maintaining performance increase for large prompts. Batch size of 128 is recommended to be used.  The patch eliminates caching of generated tokens and quantizes the length of the cache to batch size, so it requires re-processing of both generated tokens and the last partial batch in all prompts to guarantee deterministic results.